Ensured that even if the arch-setup fails, the whole install script will not fail
Added `||true` to the end of the arch-setup.sh calling part, so that a
failure there does not result in a failure of the rest of the setup
script

+++ install/emulator/emulate-arch-live-iso-install.sh
@@ -17,28 +17,40 @@ else
   echo "Compiled custom arch ISO found locally, loading existing."
 fi
 
-# Configure disk
-disk_name="arch-disk.img"
+####################
+## Qemu Paths ######
+####################
+
 emulator_dir="$dotfiles/install/emulator"
+
+disk_name="arch-disk.img"
 disk_dir="$emulator_dir/disk"
 disk_path="$disk_dir/$disk_name"
+
+nvram_dir="$emulator_dir/nvram"
+nvram_name="OVFM_VARS.fd"
+nvram_path="$nvram_dir/$nvram_name"
+
 mkdir -p "$disk_dir"
 
-echo "Recreating new empty virtual disk..."
-rm -f "$disk_path"
-qemu-img create -f qcow2 "$disk_path" 20G
+read -p "Do you want to recreate the virtual disk and NVRAM file? [y/N]: " recreate_choice
 
-# Check if there's an existing nvram file
-nvram_path="$emulator_dir/nvram"
-nvram_name="OVFM_VARS.fd"
-echo "Recreating new empty nvram file..."
-rm -f "$nvram_path/$nvram_name"
-mkdir -p "$nvram_path"
-cp /usr/share/edk2/x64/OVMF_VARS.4m.fd "$nvram_path/$nvram_name"
+if [[ "$recreate_choice" =~ ^[Yy]$ ]]; then
+  echo "Recreating new empty virtual disk..."
+  rm -f "$disk_path"
+  qemu-img create -f qcow2 "$disk_path" 20G
+
+  echo "Recreating new empty NVRAM file..."
+  rm -f "$nvram_path"
+  mkdir -p "$nvram_dir"
+  cp /usr/share/edk2/x64/OVMF_VARS.4m.fd "$nvram_path"
+else
+  echo "Keeping existing virtual disk and NVRAM file."
+fi
 
 qemu-system-x86_64 \
   -drive if=pflash,format=raw,readonly=on,file=/usr/share/edk2/x64/OVMF_CODE.4m.fd \
-  -drive if=pflash,format=raw,file="$nvram_path/$nvram_name" \
+  -drive if=pflash,format=raw,file="$nvram_path" \
   -enable-kvm \
   -cpu host \
   -m 2048 \
@@ -46,12 +58,12 @@ qemu-system-x86_64 \
   -drive file=$disk_path,format=qcow2,if=virtio \
   -cdrom $iso_path \
   -virtfs local,id=shared,path="$dotfiles/install/iso/overlay/airootfs/root",security_model=mapped,mount_tag=hostshare \
-  -boot menu=on \
+  -boot order=d \
   -nographic
 
 # For dynamic file editing / debugging in the VM, before I bake the script into the ISO itself:
-# mkdir /mnt/hostshare
-# mount -t 9p -o trans=virtio,version=9p2000.L hostshare /mnt/hostshare
+# mkdir /tmp/hostshare
+# mount -t 9p -o trans=virtio,version=9p2000.L hostshare /tmp/hostshare
 
 # In my .nvimrc, see how I can enable line-wraparound by default...
 # Also, as a matter of urgency (next thing bc is actually useful), make
